Adding Retry

+++ main.py
@@ -21,14 +21,14 @@ print('num of arguments: ', len(sys.argv))
 input_args = sys.argv
 
 list_of_subreddits = ['hentai','Futanari','FutanariGifs','futanari_Comics','FutanariHentai','blowjobsandwich',
-                    'MikeAdriano','GirlsFinishingTheJob','cumsluts','javdreams','nsfwcosplay', 'deepthroat',
-                     'bukkake','PornIsCheating','pornrelapsed','bigtitsinbikinis','ActuallyHugeCumshots','SpitRoasted',
-                     'ClassyPornstars','ModelsGoneMild','tiktokthots','Exxxtras','Oilporn','HENTAI_GIF','FemboyHentai']
+					'MikeAdriano','GirlsFinishingTheJob','cumsluts','javdreams','nsfwcosplay', 'deepthroat',
+					 'bukkake','PornIsCheating','pornrelapsed','bigtitsinbikinis','ActuallyHugeCumshots','SpitRoasted',
+					 'ClassyPornstars','ModelsGoneMild','tiktokthots','Exxxtras','Oilporn','HENTAI_GIF','FemboyHentai','blowbang']
 star_subreddits = ['AngelaWhite','RileyReid','MiaMalkova','GabbieCarter','abelladanger',
-                   'AdrianaChechik','LenaPaul','RemyLaCroix','Sashagrey','anriokita',
-                   'GiannaMichaels','BrandiLove','sophiedee','LisaAnn','JadaStevens',
-                   'KendraLust','valentinanappi','karleegrey','MandyMuse','Evalovia','AlettaOcean',
-                   'SaraJay','NicoletteSheaNew','KristyBlack','NatashaNice']
+				   'AdrianaChechik','LenaPaul','RemyLaCroix','Sashagrey','anriokita',
+				   'GiannaMichaels','BrandiLove','sophiedee','LisaAnn','JadaStevens',
+				   'KendraLust','valentinanappi','karleegrey','MandyMuse','Evalovia','AlettaOcean',
+				   'SaraJay','NicoletteSheaNew','KristyBlack','NatashaNice']
 all_subreddits = list_of_subreddits+star_subreddits
 
 reddit = praw.Reddit(client_id=input_args[1], #REDDIT_CLIENT_ID
@@ -50,12 +50,12 @@ twitter_api_authorized = Api(
 #red_gifs_api.login()
 
 with open ('todays_list.ob', 'rb') as fp:
-    todays_alreadysent_list = pickle.load(fp)
-    print(todays_alreadysent_list)
+	todays_alreadysent_list = pickle.load(fp)
+	print(todays_alreadysent_list)
 if datetime.datetime.today in todays_alreadysent_list:
-    pass
+	pass
 else:
-    todays_alreadysent_list = [datetime.datetime.today]
+	todays_alreadysent_list = [datetime.datetime.today]
 
 filename = 'to_upload.mp4'
 
@@ -67,31 +67,48 @@ for x in reddit.subreddit(subreddit).top(time_filter='day',limit=25):
 	if 'redgifs' in x.url:
 		url = x.url
 		print(url, flush=True)
-		#resp = requests.get(url) # making requests to server
-		#with open(filename, "wb") as f: # opening a file handler to create new file 
-		#    f.write(resp.content) # writing content to file
-		#red_gifs_api.download(url, filename)
-
 		video_url = helper.get_redgifs_embedded_video_url(redgifs_url=url,output_fn=filename)
-		
 		tweet_title=str(x.title).replace('my','the').replace('I','they').replace("I'm","they're") \
-                .replace("I've","they've").replace("I'd","they'd") + ' #' +str(subreddit)
+				.replace("I've","they've").replace("I'd","they'd") + ' #' +str(subreddit)
 		#print(tweet_title, flush=True)
 		break
-tweet_title=str(x.title).replace('my','the').replace('I','they').replace("I'm","they're") \
-                .replace("I've","they've").replace("I'd","they'd") + ' #' +str(subreddit)
-
 todays_alreadysent_list.append(subreddit)
-
 with open('todays_list.ob', 'wb') as fp:
-    #pickle.dump([], fp)
-    pickle.dump(todays_alreadysent_list, fp)
+	#pickle.dump([], fp)
+	pickle.dump(todays_alreadysent_list, fp)
+
+if not os.path.isfile(filename):
+	subreddits_to_choose_from = [x for x in all_subreddits if x not in todays_alreadysent_list]
+	subreddit = random.choice(subreddits_to_choose_from)
+	print(subreddit)
+	for x in reddit.subreddit(subreddit).top(time_filter='day',limit=25):
+		print(x.url)
+		if 'redgifs' in x.url:
+			url = x.url
+			print(url, flush=True)
+			video_url = helper.get_redgifs_embedded_video_url(redgifs_url=url,output_fn=filename)
+			tweet_title=str(x.title).replace('my','the').replace('I','they').replace("I'm","they're") \
+					.replace("I've","they've").replace("I'd","they'd") + ' #' +str(subreddit)
+			#print(tweet_title, flush=True)
+			break
+	todays_alreadysent_list.append(subreddit)
+	with open('todays_list.ob', 'wb') as fp:
+		#pickle.dump([], fp)
+		pickle.dump(todays_alreadysent_list, fp)
+	
+else:
+	pass
+
+
+
+tweet_title=str(x.title).replace('my','the').replace('I','they').replace("I'm","they're") \
+				.replace("I've","they've").replace("I'd","they'd") + ' #' +str(subreddit)
 
 total_bytes = os.path.getsize(filename)
 print(total_bytes)
 resp = twitter_api_authorized.upload_media_chunked_init(
-    total_bytes=total_bytes,
-    media_type="video/mp4",
+	total_bytes=total_bytes,
+	media_type="video/mp4",
 )
 media_id = resp.media_id_string
 #print(media_id)
@@ -101,16 +118,16 @@ bytes_sent = 0
 file = open(filename, 'rb')
 idx=0
 while bytes_sent < total_bytes:
-    chunk = file.read(4*1024*1024)
-    status = twitter_api_authorized.upload_media_chunked_append(
-            media_id=media_id,
-            media=chunk,
-            segment_index=idx
-        )
-    idx = idx+1
-    
-    bytes_sent = file.tell()
-    print(idx, media_id, status, bytes_sent)
+	chunk = file.read(4*1024*1024)
+	status = twitter_api_authorized.upload_media_chunked_append(
+			media_id=media_id,
+			media=chunk,
+			segment_index=idx
+		)
+	idx = idx+1
+	
+	bytes_sent = file.tell()
+	print(idx, media_id, status, bytes_sent)
 
 resp = twitter_api_authorized.upload_media_chunked_finalize(media_id=media_id)
 print(resp)
